Show live match count on home screen
