+++ app/src/main/java/com/besosn/app/presentation/ui/home/HomeFragment.kt
@@ -6,6 +6,7 @@ import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
 import com.besosn.app.R
 import com.besosn.app.databinding.FragmentHomeBinding
+import com.besosn.app.presentation.ui.matches.MatchesLocalDataSource
 
 class HomeFragment : Fragment(R.layout.fragment_home) {
 
@@ -31,10 +32,30 @@ class HomeFragment : Fragment(R.layout.fragment_home) {
         binding.btnSettings.setOnClickListener {
             findNavController().navigate(R.id.action_homeFragment_to_settingsFragment)
         }
+
+        updateMatchesCount()
+    }
+
+    override fun onResume() {
+        super.onResume()
+        if (_binding != null) {
+            updateMatchesCount()
+        }
     }
 
     override fun onDestroyView() {
         super.onDestroyView()
         _binding = null
+    }
+
+    private fun updateMatchesCount() {
+        val context = context ?: return
+        val totalMatches = MatchesLocalDataSource.loadMatches(context).size
+        val countText = resources.getQuantityString(
+            R.plurals.home_matches_count,
+            totalMatches,
+            totalMatches,
+        )
+        binding.matchesCount.text = countText
     }
 }
